refactor(utils): clean up Logger, turn into one class
