+++ packages/utils/src/Logger.ts
@@ -1,8 +1,22 @@
 import pino from 'pino'
 import path from 'path'
 import _ from 'lodash'
+import pinoPretty from 'pino-pretty'
 
-export class LoggerCommon {
+const parseBoolean = (value: string|undefined) => {
+    switch (value) {
+        case 'true':
+            return true
+        case 'false':
+            return false
+        case undefined:
+            return undefined
+        default:
+            throw new Error('Invalid boolean value: ${value}')
+    }
+}
+
+export class Logger {
     static NAME_LENGTH = 20
 
     static createName(module: NodeJS.Module, context?: string): string {
@@ -20,7 +34,23 @@ export class LoggerCommon {
 
     protected readonly logger: pino.Logger
 
-    constructor(options: pino.LoggerOptions, destinationStream?: { write(msg: string): void }, ) {
+    constructor(module: NodeJS.Module, context?: string, destinationStream?: { write(msg: string): void }) {
+        const options: pino.LoggerOptions = {
+            name: Logger.createName(module, context),
+            enabled: !process.env.NOLOG,
+            level: process.env.LOG_LEVEL || 'info',
+            // explicitly pass prettifier, otherwise pino may try to lazy require it,
+            // which can fail when under jest+typescript, due to some CJS/ESM
+            // incompatibility leading to throwing an error like:
+            // "prettyFactory is not a function"
+            prettifier: process.env.NODE_ENV === 'production' ? undefined : pinoPretty,
+            prettyPrint: process.env.NODE_ENV === 'production' ? false : {
+                colorize: parseBoolean(process.env.LOG_COLORS) ?? true,
+                translateTime: 'yyyy-mm-dd"T"HH:MM:ss.l',
+                ignore: 'pid,hostname',
+                levelFirst: true,
+            }
+        }
         this.logger = (destinationStream !== undefined) ? pino(options, destinationStream) : pino(options)
     }
