+++ packages/utils/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./tsconfig.jest.json"
+}
